backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Domain Generator Dependencies
duckduckgo-search==6.3.5
httpx==0.27.2
beautifulsoup4==4.12.3
selenium==4.26.1
requests==2.32.3
